refactor(unstyled): one file per component, matching mantine
chrome.tsx and tables.tsx split into ActiveFilterChips, BulkActionBar,
PaginationFooter, ErrorState, TableSkeleton, ExpandToggle, RowActionButtons,
DesktopTable and MobileCards. No behaviour, exports or public API change.
